Move memo and PureComponent into their own modules (+0 B)
