+++ compat/src/index.js
@@ -26,6 +26,8 @@ import { assign } from './util';
 import { Suspense, lazy } from './suspense';
 import { SuspenseList } from './suspense-list';
 import { createPortal } from './createPortal';
+import { memo } from './memo';
+import { PureComponent } from './PureComponent';
 import { Children } from './Children';
 
 const version = '16.8.0'; // trick libraries to think we are react
@@ -148,18 +150,6 @@ function unmountComponentAtNode(container) {
 }
 
 /**
- * Check if two objects have a different shape
- * @param {object} a
- * @param {object} b
- * @returns {boolean}
- */
-function shallowDiffers(a, b) {
-	for (let i in a) if (i !== '__source' && !(i in b)) return true;
-	for (let i in b) if (i !== '__source' && a[i] !== b[i]) return true;
-	return false;
-}
-
-/**
  * Get the matching DOM node for a component
  * @param {import('./internal').Component} component
  * @returns {import('./internal').PreactElement | null}
@@ -172,57 +162,8 @@ function findDOMNode(component) {
 	);
 }
 
-/**
- * Component class with a predefined `shouldComponentUpdate` implementation
- */
-class PureComponent extends Component {
-	constructor(props) {
-		super(props);
-		// Some third-party libraries check if this property is present
-		this.isPureReactComponent = true;
-	}
-
-	shouldComponentUpdate(props, state) {
-		return (
-			shallowDiffers(this.props, props) || shallowDiffers(this.state, state)
-		);
-	}
-}
-
 // Some libraries like `react-virtualized` explicitly check for this.
 Component.prototype.isReactComponent = {};
-
-/**
- * Memoize a component, so that it only updates when the props actually have
- * changed. This was previously known as `React.pure`.
- * @param {import('./internal').FunctionalComponent} c functional component
- * @param {(prev: object, next: object) => boolean} [comparer] Custom equality function
- * @returns {import('./internal').FunctionalComponent}
- */
-function memo(c, comparer) {
-	function shouldUpdate(nextProps) {
-		let ref = this.props.ref;
-		let updateRef = ref == nextProps.ref;
-		if (!updateRef && ref) {
-			ref.call ? ref(null) : (ref.current = null);
-		}
-
-		if (!comparer) {
-			return shallowDiffers(this.props, nextProps);
-		}
-
-		return !comparer(this.props, nextProps) || !updateRef;
-	}
-
-	function Memoed(props) {
-		this.shouldComponentUpdate = shouldUpdate;
-		return h(c, assign({}, props));
-	}
-	Memoed.prototype.isReactComponent = true;
-	Memoed.displayName = 'Memo(' + (c.displayName || c.name) + ')';
-	Memoed._forwarded = true;
-	return Memoed;
-}
 
 /**
  * Pass ref down to a child. This is mainly used in libraries with HOCs that
